fix: Status code was not accessible from typescript #284
