+++ sdk/src/reply.ts
@@ -16,6 +16,9 @@
 
 import { Method } from 'protobufjs';
 import { Metadata } from './metadata';
+import { GrpcStatus } from './akkaserverless';
+
+// Note that there is a parallel reply.jsdoc that needs to be changed for API changes here to be visible
 
 /**
  * Side effect for a reply.
@@ -265,27 +268,4 @@ export class Failure {
   getStatus(): GrpcStatus | undefined {
     return this.status;
   }
-}
-
-/**
- * The GRPC status codes.
- */
-export enum GrpcStatus {
-  Ok = 0,
-  Cancelled = 1,
-  Unknown = 2,
-  InvalidArgument = 3,
-  DeadlineExceeded = 4,
-  NotFound = 5,
-  AlreadyExists = 6,
-  PermissionDenied = 7,
-  ResourceExhausted = 8,
-  FailedPrecondition = 9,
-  Aborted = 10,
-  OutOfRange = 11,
-  Unimplemented = 12,
-  Internal = 13,
-  Unavailable = 14,
-  DataLoss = 15,
-  Unauthenticated = 16,
 }
